% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgestream.R
\name{avg_clustering}
\alias{avg_clustering}
\title{Average local clustering coefficient}
\usage{
avg_clustering(G)
}
\arguments{
\item{G}{A StreamGraph.}
}
\value{
Numeric in [0, 1].
}
\description{
Mean of [clustering_coefficient()] over all touched nodes.
}
\examples{
G <- stream_graph()
for (e in list(c(0, 1), c(1, 2), c(0, 2), c(2, 3))) add_edge(G, e[1], e[2])
avg_clustering(G)
}
