% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgestream.R
\name{avg_degree}
\alias{avg_degree}
\title{Average degree}
\usage{
avg_degree(G)
}
\arguments{
\item{G}{A StreamGraph.}
}
\value{
Numeric: 2m/n undirected, m/n directed.
}
\description{
Average degree
}
\examples{
G <- stream_graph(); for (i in 0:3) add_edge(G, i, i + 1L)
avg_degree(G)
}
