% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgestream.R
\name{clear_graph}
\alias{clear_graph}
\title{Clear a graph while retaining allocated storage}
\usage{
clear_graph(G)
}
\arguments{
\item{G}{A StreamGraph.}
}
\value{
Invisibly, G.
}
\description{
Clear a graph while retaining allocated storage
}
