% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamgraph.R
\name{component_nodes}
\alias{component_nodes}
\title{All touched nodes in a node's component}
\usage{
component_nodes(G, u)
}
\arguments{
\item{G}{A StreamGraph.}

\item{u}{Node ID (0-indexed).}
}
\value{
Integer vector of node IDs (0-indexed).
}
\description{
All touched nodes in a node's component
}
\examples{
G <- stream_graph(); add_edge(G, 0L, 1L); add_edge(G, 1L, 2L)
component_nodes(G, 0L)
}
