% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamgraph.R
\name{component_size}
\alias{component_size}
\title{Number of nodes in a node's component}
\usage{
component_size(G, u)
}
\arguments{
\item{G}{A StreamGraph.}

\item{u}{Node ID (0-indexed).}
}
\value{
Integer.
}
\description{
Number of nodes in a node's component
}
\examples{
G <- stream_graph(); for (i in 0:4) add_edge(G, i, i + 1L)
component_size(G, 0L)
}
