% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamgraph.R
\name{degree}
\alias{degree}
\title{Degree of a node}
\usage{
degree(G, u)
}
\arguments{
\item{G}{A StreamGraph.}

\item{u}{Node ID (0-indexed).}
}
\value{
Integer.
}
\description{
Degree of a node
}
\examples{
G <- stream_graph(); add_edge(G, 0L, 1L); add_edge(G, 0L, 2L)
degree(G, 0L)
}
