% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgestream.R
\name{density}
\alias{density}
\title{Edge density}
\usage{
density(G)
}
\arguments{
\item{G}{A StreamGraph.}
}
\value{
Numeric: 2m/(n(n-1)) undirected, m/(n(n-1)) directed.
}
\description{
Edge density
}
\examples{
G <- stream_graph(); for (i in 0:3) add_edge(G, i, i + 1L)
density(G)
}
