% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgestream.R
\docType{package}
\name{edgestream-package}
\alias{edgestream}
\alias{edgestream-package}
\title{edgestream: streaming (dynamic) graph analytics}
\description{
Maintains connected components, triangle counts, degrees and basic centrality
incrementally as edges arrive. The graph object is an external pointer; node
IDs are non-negative integers, 0-indexed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bagol1000/edgestream}
  \item Report bugs at \url{https://github.com/bagol1000/edgestream/issues}
}

}
\author{
\strong{Maintainer}: bagol1000 \email{01199218@pw.edu.pl}

Authors:
\itemize{
  \item bagol1000 \email{01199218@pw.edu.pl}
}

}
\keyword{internal}
