% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgestream.R
\name{has_edge}
\alias{has_edge}
\title{Whether an edge exists}
\usage{
has_edge(G, u, v)
}
\arguments{
\item{G}{A StreamGraph.}

\item{u, v}{Node IDs (0-indexed).}
}
\value{
Logical.
}
\description{
Whether an edge exists
}
\examples{
G <- stream_graph(); add_edge(G, 0L, 1L)
has_edge(G, 0L, 1L)
}
