% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamgraph.R
\name{in_degree}
\alias{in_degree}
\title{In-degree of a node}
\usage{
in_degree(G, u)
}
\arguments{
\item{G}{A StreamGraph.}

\item{u}{Node ID (0-indexed).}
}
\value{
Integer.
}
\description{
Equals [degree()] in undirected graphs.
}
\examples{
G <- stream_graph(directed = TRUE); add_edge(G, 0L, 2L); add_edge(G, 1L, 2L)
in_degree(G, 2L)  # 2
degree(G, 2L)     # 0 (out-degree)
}
