% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgestream.R
\name{in_neighbours}
\alias{in_neighbours}
\title{Sorted in-neighbours of a node}
\usage{
in_neighbours(G, u)
}
\arguments{
\item{G}{A StreamGraph.}

\item{u}{Node ID (0-indexed).}
}
\value{
Sorted integer vector of in-neighbour IDs (0-indexed).
}
\description{
Equals [neighbours()] in undirected graphs.
}
\examples{
G <- stream_graph(directed = TRUE); add_edge(G, 0L, 2L); add_edge(G, 1L, 2L)
in_neighbours(G, 2L)  # c(0L, 1L)
}
