% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamgraph.R
\name{local_triangles}
\alias{local_triangles}
\title{Triangles incident to a node}
\usage{
local_triangles(G, u)
}
\arguments{
\item{G}{A StreamGraph.}

\item{u}{Node ID (0-indexed).}
}
\value{
Integer.
}
\description{
Triangles incident to a node
}
\examples{
G <- stream_graph()
for (e in list(c(0,1), c(1,2), c(0,2))) add_edge(G, e[1], e[2])
local_triangles(G, 0L)
}
