% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgestream.R
\name{max_degree}
\alias{max_degree}
\title{Maximum degree}
\usage{
max_degree(G)
}
\arguments{
\item{G}{A StreamGraph.}
}
\value{
Integer.
}
\description{
Maintained incrementally, O(1) to query. In directed graphs this is the
maximum out-degree.
}
\examples{
G <- stream_graph(); for (i in 0:3) add_edge(G, i, i + 1L)
max_degree(G)
}
