% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgestream.R
\name{n_edges}
\alias{n_edges}
\title{Number of edges}
\usage{
n_edges(G)
}
\arguments{
\item{G}{A StreamGraph.}
}
\value{
Numeric.
}
\description{
Number of edges
}
\examples{
G <- stream_graph(); add_edge(G, 0L, 1L)
n_edges(G)
}
