% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgestream.R
\name{n_ids}
\alias{n_ids}
\title{Size of the allocated node-ID range}
\usage{
n_ids(G)
}
\arguments{
\item{G}{A StreamGraph.}
}
\value{
Numeric, equal to max current node ID plus one.
}
\description{
Size of the allocated node-ID range
}
