% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgestream.R
\name{n_nodes}
\alias{n_nodes}
\title{Number of touched (referenced) nodes}
\usage{
n_nodes(G)
}
\arguments{
\item{G}{A StreamGraph.}
}
\value{
Integer.
}
\description{
Number of touched (referenced) nodes
}
\examples{
G <- stream_graph(); add_edge(G, 0L, 1L)
n_nodes(G)
}
