% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgestream.R
\name{n_strong_components}
\alias{n_strong_components}
\title{Number of strongly connected components}
\usage{
n_strong_components(G)
}
\arguments{
\item{G}{A StreamGraph.}
}
\value{
Integer.
}
\description{
Number of strongly connected components
}
\examples{
D <- stream_graph(directed = TRUE)
add_edge(D, 0L, 1L); add_edge(D, 1L, 0L); add_edge(D, 1L, 2L)
n_strong_components(D)  # 2
}
