% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamgraph.R
\name{neighbours}
\alias{neighbours}
\title{Sorted neighbours of a node}
\usage{
neighbours(G, u)
}
\arguments{
\item{G}{A StreamGraph.}

\item{u}{Node ID (0-indexed).}
}
\value{
Sorted integer vector of neighbour IDs (0-indexed).
}
\description{
In directed graphs these are the out-neighbours; see [in_neighbours()].
}
\examples{
G <- stream_graph(); add_edge(G, 5L, 2L); add_edge(G, 5L, 1L)
neighbours(G, 5L)
}
