% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgestream.R
\name{reserve_nodes}
\alias{reserve_nodes}
\alias{reserve_edges}
\title{Reserve graph storage}
\usage{
reserve_nodes(G, n)

reserve_edges(G, m)
}
\arguments{
\item{G}{A StreamGraph.}

\item{n}{Size of the reserved node-ID range.}

\item{m}{Approximate number of distinct edges.}
}
\description{
Reserve graph storage
}
