% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgestream.R
\name{save_graph}
\alias{save_graph}
\title{Save a graph to a binary file}
\usage{
save_graph(G, path)
}
\arguments{
\item{G}{A StreamGraph.}

\item{path}{Output file path.}
}
\value{
Invisibly NULL.
}
\description{
Save a graph to a binary file
}
\examples{
G <- stream_graph(); add_edge(G, 0L, 1L)
p <- tempfile(fileext = ".esg")
save_graph(G, p)
}
