% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamgraph.R
\docType{package}
\name{streamgraph-package}
\alias{streamgraph}
\alias{streamgraph-package}
\title{streamgraph: streaming (dynamic) graph analytics}
\description{
Maintains connected components, triangle counts, degrees and basic centrality
incrementally as edges arrive. The graph object is an external pointer; node
IDs are non-negative integers, 0-indexed.
}
\author{
\strong{Maintainer}: bagol1000 \email{01199218@pw.edu.pl}

Authors:
\itemize{
  \item bagol1000 \email{01199218@pw.edu.pl}
}

}
\keyword{internal}
