% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamgraph.R
\name{triangle_count}
\alias{triangle_count}
\title{Global triangle count}
\usage{
triangle_count(G)
}
\arguments{
\item{G}{A StreamGraph.}
}
\value{
Numeric; each triangle counted once.
}
\description{
Global triangle count
}
\examples{
G <- stream_graph()
for (e in list(c(0,1), c(1,2), c(0,2))) add_edge(G, e[1], e[2])
triangle_count(G)
}
