% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgestream.R
\name{window_advance}
\alias{window_advance}
\title{Advance a sliding window}
\usage{
window_advance(x, timestamp)
}
\arguments{
\item{x}{A window from [sliding_window_graph()].}

\item{timestamp}{Finite timestamp, not smaller than the preceding timestamp.}
}
\value{
Number of expired edges.
}
\description{
Advance a sliding window
}
