% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgestream.R
\name{degree_histogram}
\alias{degree_histogram}
\title{Degree histogram}
\usage{
degree_histogram(G)
}
\arguments{
\item{G}{A StreamGraph.}
}
\value{
Named numeric vector; entry "d" is the number of nodes of degree d.
}
\description{
In directed graphs the histogram is over out-degrees.
}
\examples{
G <- stream_graph(); add_edge(G, 0L, 1L); add_edge(G, 0L, 2L)
degree_histogram(G)
}
